removed nonsense warning
the "asset dependencies" in the "~/Documents/musik" folder will only ever be non-vital. Not having the asset library won't cause musik to crash, but it will definitely feel empty. I updated the warnings to reflect that—still prompting you to create a library on start, but not requiring it. I will likely have the popup remember to stay hidden when pressing "no" once I create a settings system that saves to storage. In the meantime, a temporary fix is to either just let it create the library or put a random file/folder in the "~/Documents/musik" location to satisfy the os.path.exists() condition and prevent the rest of initialize_musik_library() from running.