temporarily remove uke-lick0.mp3
uke-lick0.mp3 is too long to use it in mpg123 comfortably (it plays for super long and keep playing even when the app is force killed because it's a separate mpg123 process). So I've temporarily removed its binding in the default pads. I have left the sample in the asset library, but it will not be used by default until we migrate away from mpg123 and the one-shot can be elegantly interrupted